// packages/react-cv/src/components/resume-types.ts
export interface Location {
	// To add multiple address lines, use \n. For example, 1234 Glücklichkeit Straße\nHinterhaus 5. Etage li.
	address?: string;

	// Current address' postal code, format accoridng to your country
	postalCode?: string;

	// The address' city
	city: string;

	// code as per ISO-3166-1 ALPHA-2, e.g. US, AU, IN
	countryCode: string;

	// The general region where you live. Can be a US state, or a province, for instance.
	region: string;
}

export interface Profile {
	// e.g. Facebook or Twitter
	network: string;

	// e.g. neutralthoughts
	username: string;

	// e.g. http://twitter.example.com/neutralthoughts
	url: string;
}

export interface Basic {
	// Your name
	name: string;

	// Current description, e.g. Web Developer
	label?: string;

	// URL (as per RFC 3986) to a image in JPEG or PNG format
	image?: string;

	// e.g. thomas@gmail.com
	email?: string;

	// Phone numbers are stored as strings so use any format you like, e.g. 712-117-2923
	phone?: string;

	// URL (as per RFC 3986) to your website, e.g. personal homepage
	url?: string;

	// Write a short 2-3 sentence biography about yourself
	summary?: string;

	// Your current location information, sensitive
	location?: Location;

	// Specify any number of social networks that you participate in
	profiles?: Profile[];
}

export interface Work {
	// e.g. Facebook
	company: string;

	// e.g. Menlo Park, CA
	location?: string;

	// e.g. Social Media Company
	description?: string;

	// e.g. Software Engineer
	position?: string;

	// e.g. http://facebook.example.com
	url?: string;

	// resume.json uses the ISO 8601 date standard e.g. 2014-06-29
	startDate?: string;

	// e.g. 2012-06-29
	endDate?: string;

	// Give an overview of your responsibilities at the company
	summary?: string;

	// Specify multiple accomplishments, e.g. Increased profits by 20% from 2011-2012 through viral advertising
	highlights?: string[];

	// Keywords or skills, e.g. React, JavaScript
	keywords?: string[];
}
// ...
export interface Education {
	// e.g. Massachusetts Institute of Technology
	institution: string;

	// e.g. Arts
	area?: string;

	// e.g. Bachelor
	studyType?: string;

	// resume.json uses the ISO 8601 date standard e.g. 2014-06-29
	startDate?: string;

	// e.g. 2012-06-29
	endDate?: string;

	// grade point average, e.g. 3.67/4.0
	gpa?: string;

	// List notable courses/subjects, e.g. H1302 - Introduction to American history
	courses?: string[];
}
// ...
export interface Skill {
	id?: string;
	// e.g. Web Development
	name: string;

	// "e.g. Master
	level?: string;

	// List some keywords pertaining to this skill, e.g. HTML
	keywords?: string[];
	preferred?: string[]; // Preferred skills
}